refactor: フロントエンド ページ分割・共通化
- SettingsPage (812行→62行) を4サブコンポーネントに分割
  - ProfileSection / SkillsSection / IntegrationSection / AccountSection
- useSettings フック新設で全ロジックをページから分離
- ChatPage (487行→349行) のロジックを useChat フックに抽出
- constants/skills.ts に LANGUAGES/FRAMEWORKS を集約（3ファイルの重複解消）
- constants/styles.ts に共通 inputClass を定義

Closes #156
