+++ frontend/src/constants/skills.ts
@@ -1,0 +1,13 @@
+// skillicons.dev supported icons
+export const LANGUAGES = [
+  'java', 'typescript', 'javascript', 'python', 'go', 'rust', 'cpp', 'c', 'cs',
+  'php', 'ruby', 'swift', 'kotlin', 'scala', 'elixir', 'haskell', 'lua', 'perl',
+  'r', 'dart', 'html', 'css', 'sass', 'bash', 'powershell', 'sql', 'graphql'
+] as const;
+
+export const FRAMEWORKS = [
+  'react', 'vue', 'angular', 'svelte', 'nextjs', 'nuxtjs', 'gatsby', 'astro',
+  'spring', 'django', 'flask', 'fastapi', 'rails', 'laravel', 'express', 'nestjs',
+  'gin', 'fiber', 'actix', 'rocket', 'tailwind', 'bootstrap', 'materialui',
+  'prisma', 'graphql', 'apollo', 'redux', 'nodejs', 'deno', 'bun'
+] as const;
