+++ frontend/src/constants/styles.ts
@@ -1,0 +1,2 @@
+export const inputClass =
+  'w-full px-3 py-2 bg-gray-800/50 border border-gray-700 rounded-lg text-white text-sm placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-shadow';
