+++ frontend/src/hooks/index.ts
@@ -22,3 +22,5 @@ export { useAdvice, useAIChat, useConversations } from './useAdvice';
 export { usePomodoroTimer } from './usePomodoroTimer';
 export { useMyLevel, useLevel, useLevelBreakdown } from './useLevel';
 export { useHeatmap, useCategoryBreakdown, useProductivityScore, useWeeklyTrends, useInsights } from './useAnalytics';
+export { useSettings } from './useSettings';
+export { useChat } from './useChat';
