+++ frontend/src/pages/OnboardingPage.tsx
@@ -9,19 +9,7 @@ import { connectZenn } from '../api/zenn';
 import { connectQiita } from '../api/qiita';
 import { connectAtCoder } from '../api/atcoder';
 import toast from 'react-hot-toast';
-
-const LANGUAGES = [
-  'java', 'typescript', 'javascript', 'python', 'go', 'rust', 'cpp', 'c', 'cs',
-  'php', 'ruby', 'swift', 'kotlin', 'scala', 'elixir', 'haskell', 'lua', 'perl',
-  'r', 'dart', 'html', 'css', 'sass', 'bash', 'powershell', 'sql', 'graphql'
-];
-
-const FRAMEWORKS = [
-  'react', 'vue', 'angular', 'svelte', 'nextjs', 'nuxtjs', 'gatsby', 'astro',
-  'spring', 'django', 'flask', 'fastapi', 'rails', 'laravel', 'express', 'nestjs',
-  'gin', 'fiber', 'actix', 'rocket', 'tailwind', 'bootstrap', 'materialui',
-  'prisma', 'graphql', 'apollo', 'redux', 'nodejs', 'deno', 'bun'
-];
+import { LANGUAGES, FRAMEWORKS } from '../constants/skills';
 
 const STEPS = [
   { id: 1, icon: User },
